backend/requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
supabase==2.22.1
python-dotenv==1.1.1
pydantic==2.12.3
email-validator==2.1.1
